refactor(settings): extract shared settings UI components
